requirements.txt
bcrypt==4.1.3
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
mutagen==1.47.0
oauthlib==3.2.2
paramiko==3.4.0
pexpect==4.9.0
proto-plus==1.24.0
protobuf==5.27.2
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
spotipy==2.24.0
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
youtube-dl==2021.12.17
yt-dlp==2024.7.9
